requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-ckeditor-5==0.2.17
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
